+++ src/main/resources/db/migration/V0.1__Create_Table_Goods.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `t_goods` (
-  `id` INT NOT NULL PRIMARY KEY ,
+  `id` BIGINT(20) NOT NULL PRIMARY KEY ,
   `name` VARCHAR(255) NOT NULL ,
   `description` VARCHAR(255) NOT NULL ,
   `price` DOUBLE DEFAULT 0,
-  `amount` INT DEFAULT 0
+  `amount` BIGINT(20) DEFAULT 0
 )
